app/requirements.txt
psycopg2-binary
soda-core-postgres
dotenv
